% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_api.R
\name{artifact_path}
\alias{artifact_path}
\title{Path to a baked artifact file under \verb{inst/report/artifacts/}}
\usage{
artifact_path(id, ext)
}
\arguments{
\item{id}{Replication id.}

\item{ext}{File extension (\code{png}, \code{html}, \code{rds}, ...).}
}
\description{
Path to a baked artifact file under \verb{inst/report/artifacts/}
}
\keyword{internal}
