% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{deal_labels}
\alias{deal_labels}
\title{Deal labels for wave 4 vignettes}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
deal_labels
}
\description{
Deal labels for wave 4 vignettes
}
\keyword{datasets}
