% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_marginal_effects_by}
\alias{fit_marginal_effects_by}
\title{Fit marginal-effects models split by a grouping variable}
\usage{
fit_marginal_effects_by(data, split_var)
}
\arguments{
\item{data}{Wave 4 conjoint data.}

\item{split_var}{Character column name to split on.}
}
\value{
Tibble with an extra grouping column.
}
\description{
Fit marginal-effects models split by a grouping variable
}
