% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_outputs.R
\name{format_figure_1}
\alias{format_figure_1}
\alias{format_fig_1}
\alias{format_figure_2}
\alias{format_fig_2}
\alias{format_figure_3}
\alias{format_fig_3}
\alias{format_figure_4}
\alias{format_fig_4}
\alias{format_figure_5}
\alias{format_fig_5}
\alias{format_figure_6}
\alias{format_fig_6}
\alias{format_figure_7}
\alias{format_fig_7}
\alias{format_figure_8}
\alias{format_fig_8}
\title{Format a figure for display}
\usage{
format_figure_1(object, ...)

format_fig_1(object, ...)

format_figure_2(object, ...)

format_fig_2(object, ...)

format_figure_3(object, ...)

format_fig_3(object, ...)

format_figure_4(object, ...)

format_fig_4(object, ...)

format_figure_5(object, ...)

format_fig_5(object, ...)

format_figure_6(object, ...)

format_fig_6(object, ...)

format_figure_7(object, ...)

format_fig_7(object, ...)

format_figure_8(object, ...)

format_fig_8(object, ...)
}
\arguments{
\item{object}{ggplot object from the corresponding \code{\link[=make_figure_1]{make_figure_1()}} (etc.).}
}
\description{
Pass-through for ggplot objects produced by \verb{make_figure_*()} (Figures 1–8).
}
