% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_outputs.R
\name{format_table_1}
\alias{format_table_1}
\alias{format_tab_1}
\title{Format Table 1 for display}
\usage{
format_table_1(object, ...)

format_tab_1(object, ...)
}
\arguments{
\item{object}{Data frame from \code{\link[=make_table_1]{make_table_1()}}.}
}
\description{
Format Table 1 for display
}
