% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_outputs.R
\name{format_table_2}
\alias{format_table_2}
\alias{format_tab_2}
\title{Format Table 2 for display}
\usage{
format_table_2(object, ...)

format_tab_2(object, ...)
}
\arguments{
\item{object}{\code{kableExtra} table from \code{\link[=make_table_2]{make_table_2()}}.}
}
\description{
Format Table 2 for display
}
