% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_api.R
\name{get_package_object}
\alias{get_package_object}
\title{Load a dataset or object from this package}
\usage{
get_package_object(name, package = pkg_name())
}
\arguments{
\item{name}{Object name.}

\item{package}{Package name.}
}
\value{
The requested object.
}
\description{
LazyData objects are not visible in the package namespace; use \code{::} access.
}
\keyword{internal}
