% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_api.R
\name{list_replications}
\alias{list_replications}
\title{List available replications defined in \code{replication.yml}}
\usage{
list_replications()
}
\value{
A list of replication metadata entries.
}
\description{
List available replications defined in \code{replication.yml}
}
