% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_api.R
\name{load_artifact}
\alias{load_artifact}
\title{Load a precomputed artifact from \verb{inst/report/artifacts/}}
\usage{
load_artifact(id)
}
\arguments{
\item{id}{Replication id.}
}
\value{
File path or loaded object.
}
\description{
Load a precomputed artifact from \verb{inst/report/artifacts/}
}
