% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_1.R
\name{make_figure_1}
\alias{make_figure_1}
\alias{make_fig_1}
\title{Figure 1 — distribution of support for contributions of different sizes}
\usage{
make_figure_1(data = wave4_conjoint)

make_fig_1(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A multi-panel ggplot object from \code{egg::ggarrange}.
}
\description{
Figure 1 — distribution of support for contributions of different sizes
}
