% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_2.R
\name{make_figure_2}
\alias{make_figure_2}
\alias{make_fig_2}
\title{Figure 2 — marginal effects of conditions}
\usage{
make_figure_2(data = wave4_conjoint)

make_fig_2(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A ggplot object.
}
\description{
Figure 2 — marginal effects of conditions
}
