% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_3.R
\name{make_figure_3}
\alias{make_figure_3}
\alias{make_fig_3}
\title{Figure 3 — effect of video treatment on individual solidarity}
\usage{
make_figure_3(data = wave2_survey)

make_fig_3(data = wave2_survey)
}
\arguments{
\item{data}{Wave 2 survey data. Defaults to \link{wave2_survey}.}
}
\value{
A ggplot object.
}
\description{
Figure 3 — effect of video treatment on individual solidarity
}
