% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_4.R
\name{make_figure_4}
\alias{make_figure_4}
\alias{make_fig_4}
\title{Figure 4 — levels of support by party}
\usage{
make_figure_4(data = wave4_conjoint)

make_fig_4(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A ggplot object.
}
\description{
Figure 4 — levels of support by party
}
