% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_6.R
\name{make_figure_6}
\alias{make_figure_6}
\alias{make_fig_6}
\title{Figure 6 — marginal effects of conditions by party}
\usage{
make_figure_6(data = wave4_conjoint)

make_fig_6(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A ggplot object.
}
\description{
Figure 6 — marginal effects of conditions by party
}
