% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_7.R
\name{make_figure_7}
\alias{make_figure_7}
\alias{make_fig_7}
\title{Figure 7 — marginal effects of conditions by migration background}
\usage{
make_figure_7(data = wave4_conjoint)

make_fig_7(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A ggplot object.
}
\description{
Figure 7 — marginal effects of conditions by migration background
}
