% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figure_8.R
\name{make_figure_8}
\alias{make_figure_8}
\alias{make_fig_8}
\title{Figure 8 — results refreshment sample}
\usage{
make_figure_8(data = wave4_conjoint)

make_fig_8(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A ggplot object.
}
\description{
Figure 8 — results refreshment sample
}
