% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_1.R
\name{make_table_1}
\alias{make_table_1}
\alias{make_tab_1}
\title{Table 1 — structural parameter estimates}
\usage{
make_table_1(data = wave4_conjoint)

make_tab_1(data = wave4_conjoint)
}
\arguments{
\item{data}{Wave 4 conjoint data. Defaults to \link{wave4_conjoint}.}
}
\value{
A data frame of parameter estimates with confidence intervals.
}
\description{
Table 1 — structural parameter estimates
}
