% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep1371journalpone0278337-package.R
\docType{package}
\name{rep1371journalpone0278337-package}
\alias{rep1371journalpone0278337}
\alias{rep1371journalpone0278337-package}
\title{rep1371journalpone0278337: Replication Package for Geissler et al. (2022) Vaccine Solidarity}
\description{
Self-contained replication package for "Public support for global vaccine sharing in the COVID-19 pandemic: Evidence from Germany" (PLOS ONE, 2022; doi:10.1371/journal.pone.0278337). Provides analysis data, preparation scripts, table and figure builders, precomputed report artifacts, and tests.
}
\author{
\strong{Maintainer}: Macartan Humphreys \email{macartan.humphreys@wzb.eu}

Authors:
\itemize{
  \item Macartan Humphreys \email{macartan.humphreys@wzb.eu}
  \item Ferdinand Geissler
  \item Felix Hartmann
  \item Heike Klüver
  \item Johannes Giesecke
}

}
\keyword{internal}
