% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_api.R
\name{replication_meta}
\alias{replication_meta}
\title{Read package replication metadata}
\usage{
replication_meta()
}
\value{
Parsed contents of \code{replication.yml}.
}
\description{
Read package replication metadata
}
