% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{structural_lik_x}
\alias{structural_lik_x}
\title{Normal likelihood at the structural optimum}
\usage{
structural_lik_x(x, sigma, a, b, c, g, k, ZT, ZR, ZY, Zy)
}
\description{
Normal likelihood at the structural optimum
}
\keyword{internal}
