% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{structural_maxx}
\alias{structural_maxx}
\title{Closed-form optimum for the structural contribution model}
\usage{
structural_maxx(a, b, c, g, k, ZT, ZR, ZY, Zy)
}
\description{
Closed-form optimum for the structural contribution model
}
\keyword{internal}
