% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_labels}
\alias{vignette_labels}
\title{Vignette attribute labels for wave 2 conjoint tasks}
\format{
A data frame keyed by \code{vignr}.
}
\source{
Geissler et al. (2022), PLOS ONE.
}
\usage{
vignette_labels
}
\description{
Vignette attribute labels for wave 2 conjoint tasks
}
\keyword{datasets}
