% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wave2_survey}
\alias{wave2_survey}
\title{Wave 2 survey data (video treatment and conjoint tasks)}
\format{
A data frame with one row per respondent in wave 2.
}
\source{
Geissler et al. (2022), PLOS ONE.
}
\usage{
wave2_survey
}
\description{
Wave 2 survey data (video treatment and conjoint tasks)
}
\keyword{datasets}
